+++ tests/ast/test.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "ast/parser.h"
+#include "rocky/ast/parser.h"
 
 /* ── pretty-printer ──────────────────────────────────────── */
